v1.07 -- Craig Schlenter pointed out the "encrypt" demo doesn't call ctr_start() correctly.  That's because as of a few releases ago          I added support to set the mode of the counter at init time       -- Fixed some "testprof" make issues       -- Added RSA keygen to the math descriptors       -- Fixed install_test target ... oops       -- made the "ranlib" program renamable useful for cross-compiling       -- Made the cipher accelerators return error codes.  :-)       -- Made CCM accept a pre-scheduled key to speed it up if you use the same key for multiple packets       -- Added "Katja" public key crypto.  It's based on the recent N = p^2q work by Katja.  I added OAEP padding          to it.  Note this code has been disabled not because it doesn't work but because it hasn't been thoroughly          analyzed.   It does carry some advantages over RSA (slightly smaller public key, faster decrypt) but also          some annoying "setup" issues like the primes are smaller which makes ECM factoring more plausible.       -- Made makefile accept a NODOCS flag to disable the requirement of tetex to install LTC for you no tetex people... all 3 of ya  :-)       -- Cleaned up rsa_export() since "zero" was handled with a SHORT_INTEGER       -- Cleaned up the LIBTEST_S definitions in both GNU makefiles.  A few minor touchups as well.       -- Made the cipher ecb encrypt/decrypt return an int as well, changed ALL dependent code to check for this.       -- der_decode_choice() would fail to mark a NULL as "used" when decoding.  Fixed       -- ecc_decrypt_key() now uses find_hash_oid() to clean up the code ;-)       -- Added mp_neg() to the math descriptors.       -- Swapped arguments for the pkcs_1_mgf1() function so the hash_idx is the first param (to be more consistent)       -- Made the math descriptors buildable when RSA has been undefined       -- ECC timing demo now capable of detecting which curves have been defined       -- Refactored the ECC code so it's easier to maintain.  (note: the form of this code hasn't really changed since I first added ECC ... :-/)       -- Updated the documentation w.r.t. ECC and the accelerators to keep it current       -- Fixed bug in ltc_init_multi() which would fail to free all allocated memory on error.       -- Fixed bug in ecc_decrypt_key() which could possibly lead to overflows (if MAXBLOCKSIZE > ECC_BUF_SIZE and you have a hash that emits MAXBLOCKSIZE bytes)       -- Added encrypt/decrypt to the DSA side (basically DH with DSA parameters)       -- Updated makefiles to remove references to the old DH object files and the ecc_sys.o crap ... clean code ahead!       -- ecc_import() now checks if the point it reads in lies on the curve (to prevent degenerative points from being used)       -- ECC code now ALWAYS uses the accelerator interface.  This allows people who use the accelerators to not have the stock          ECC point add/dbl/mul code linked in.  Yeah space savings! Rah Rah Rah.       -- Added LTC_MUTEX_* support to Yarrow and Fortuna allowing you to use respective prng_state as a global PRNG state [e.g. thread-safe] if you define one of the LTC_* defines at          build time (e.g. LTC_PTHREAD == pthreads)       -- Added PPC32 support to the rotate macros (tested on an IBM PPC 405) and LTC_FAST macros (it aint fast but it's faster than stock)       -- Added ltc_mp checks in all *_make_key() and *_import() which will help catch newbs who don't register their bignum first :-)       -- the UTCTIME type was missing from der_length_sequence() [oops, oh like you've never done that]       -- the main makefile allows you to rename the make command [e.g. MAKE=gmake gmake install] so you can build LTC on platforms where the default make command sucks [e.g. BSD]       -- Added DER flexi decoder which allows the decoding of arbitrary DER encoded packets without knowing          their structure in advance (thanks to MSVC for finding 3 bugs in it just prior to release! ... don't ask)
bk: 437d1980bKd57_SqLU3aSCL-SRT77Q

+++ src/tomcrypt/testprof/ecc_test.c
@@ -29,7 +29,7 @@ int ecc_tests (void)
 	
   DO(ecc_test ());
 
-  for (s = 0; s < (int)(sizeof(sizes)/sizeof(sizes[0])); s++) {
+  for (s = 0; s < (sizeof(sizes)/sizeof(sizes[0])); s++) {
      /* make up two keys */
      DO(ecc_make_key (&yarrow_prng, find_prng ("yarrow"), sizes[s], &usera));
      DO(ecc_make_key (&yarrow_prng, find_prng ("yarrow"), sizes[s], &userb));
@@ -133,5 +133,5 @@ int ecc_tests(void)
 #endif
 
 /* $Source: /cvs/libtom/libtomcrypt/testprof/ecc_test.c,v $ */
-/* $Revision: 1.9 $ */
-/* $Date: 2005/06/14 19:43:29 $ */
+/* $Revision: 1.10 $ */
+/* $Date: 2005/10/31 14:02:24 $ */
